add methods to handle login and register
