+++ src/App.js
@@ -2,8 +2,6 @@ import './App.css';
 import 'semantic-ui-css/semantic.min.css'
 import Form from './components/InputForms/Form';
 import LoginPage from "./components/Auth/LoginPage"
-import Login from './components/Auth/Login';
-import Register from './components/Auth/Register';
 import Analytics from './components/Charts/Analytics';
 import SettingsPage from './components/Settings/SettingsPage';
 
@@ -21,8 +19,6 @@ const App = () => {
         <main>
             <Switch>
                 <Route path="/" component={LoginPage} exact />
-                <Route path="/login" component={Login} exact />
-                <Route path="/register" component={Register} exact />
                 <ProtectedRoute path="/form" component={Form} exact />
                 <ProtectedRoute path="/analytics" component={Analytics} exact />
                 <ProtectedRoute path="/settings" component={SettingsPage} exact />
